+++ sample_consensus/include/pcl/sample_consensus/impl/sac_model_circle.hpp
@@ -116,13 +116,13 @@ pcl::SampleConsensusModelCircle2D<PointT>::getDistancesToModel (const Eigen::Vec
   for (size_t i = 0; i < indices_->size (); ++i)
     // Calculate the distance from the point to the circle as the difference between
     // dist(point,circle_origin) and circle_radius
-    distances[i] = fabsf (sqrtf (
-                                ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) *
-                                ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) +
+    distances[i] = fabsf (std::sqrt (
+                                    ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) *
+                                    ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) +
 
-                                ( input_->points[(*indices_)[i]].y - model_coefficients[1] ) *
-                                ( input_->points[(*indices_)[i]].y - model_coefficients[1] )
-                                ) - model_coefficients[2]);
+                                    ( input_->points[(*indices_)[i]].y - model_coefficients[1] ) *
+                                    ( input_->points[(*indices_)[i]].y - model_coefficients[1] )
+                                    ) - model_coefficients[2]);
 }
 
 //////////////////////////////////////////////////////////////////////////
@@ -146,13 +146,13 @@ pcl::SampleConsensusModelCircle2D<PointT>::selectWithinDistance (
   {
     // Calculate the distance from the point to the sphere as the difference between
     // dist(point,sphere_origin) and sphere_radius
-    float distance = fabsf (sqrtf (
-                                  ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) *
-                                  ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) +
+    float distance = fabsf (std::sqrt (
+                                      ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) *
+                                      ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) +
 
-                                  ( input_->points[(*indices_)[i]].y - model_coefficients[1] ) *
-                                  ( input_->points[(*indices_)[i]].y - model_coefficients[1] )
-                                  ) - model_coefficients[2]);
+                                      ( input_->points[(*indices_)[i]].y - model_coefficients[1] ) *
+                                      ( input_->points[(*indices_)[i]].y - model_coefficients[1] )
+                                      ) - model_coefficients[2]);
     if (distance < threshold)
     {
       // Returns the indices of the points whose distances are smaller than the threshold
@@ -180,13 +180,13 @@ pcl::SampleConsensusModelCircle2D<PointT>::countWithinDistance (
   {
     // Calculate the distance from the point to the sphere as the difference between
     // dist(point,sphere_origin) and sphere_radius
-    float distance = fabsf (sqrtf (
-                                  ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) *
-                                  ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) +
+    float distance = fabsf (std::sqrt (
+                                      ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) *
+                                      ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) +
 
-                                  ( input_->points[(*indices_)[i]].y - model_coefficients[1] ) *
-                                  ( input_->points[(*indices_)[i]].y - model_coefficients[1] )
-                                  ) - model_coefficients[2]);
+                                      ( input_->points[(*indices_)[i]].y - model_coefficients[1] ) *
+                                      ( input_->points[(*indices_)[i]].y - model_coefficients[1] )
+                                      ) - model_coefficients[2]);
     if (distance < threshold)
       nr_p++;
   }
@@ -261,7 +261,7 @@ pcl::SampleConsensusModelCircle2D<PointT>::projectPoints (
     {
       float dx = input_->points[inliers[i]].x - model_coefficients[0];
       float dy = input_->points[inliers[i]].y - model_coefficients[1];
-      float a = sqrtf ( (model_coefficients[2] * model_coefficients[2]) / (dx * dx + dy * dy) );
+      float a = std::sqrt ( (model_coefficients[2] * model_coefficients[2]) / (dx * dx + dy * dy) );
 
       projected_points.points[inliers[i]].x = a * dx + model_coefficients[0];
       projected_points.points[inliers[i]].y = a * dy + model_coefficients[1];
@@ -285,7 +285,7 @@ pcl::SampleConsensusModelCircle2D<PointT>::projectPoints (
     {
       float dx = input_->points[inliers[i]].x - model_coefficients[0];
       float dy = input_->points[inliers[i]].y - model_coefficients[1];
-      float a = sqrtf ( (model_coefficients[2] * model_coefficients[2]) / (dx * dx + dy * dy) );
+      float a = std::sqrt ( (model_coefficients[2] * model_coefficients[2]) / (dx * dx + dy * dy) );
 
       projected_points.points[i].x = a * dx + model_coefficients[0];
       projected_points.points[i].y = a * dy + model_coefficients[1];
@@ -308,12 +308,12 @@ pcl::SampleConsensusModelCircle2D<PointT>::doSamplesVerifyModel (
   for (std::set<int>::const_iterator it = indices.begin (); it != indices.end (); ++it)
     // Calculate the distance from the point to the sphere as the difference between
     //dist(point,sphere_origin) and sphere_radius
-    if (fabsf (sqrtf (
-                     ( input_->points[*it].x - model_coefficients[0] ) *
-                     ( input_->points[*it].x - model_coefficients[0] ) +
-                     ( input_->points[*it].y - model_coefficients[1] ) *
-                     ( input_->points[*it].y - model_coefficients[1] )
-                     ) - model_coefficients[2]) > threshold)
+    if (fabsf (std::sqrt (
+                         ( input_->points[*it].x - model_coefficients[0] ) *
+                         ( input_->points[*it].x - model_coefficients[0] ) +
+                         ( input_->points[*it].y - model_coefficients[1] ) *
+                         ( input_->points[*it].y - model_coefficients[1] )
+                         ) - model_coefficients[2]) > threshold)
       return (false);
 
   return (true);
