+++ sample_consensus/include/pcl/sample_consensus/impl/sac_model_sphere.hpp
@@ -110,10 +110,9 @@ pcl::SampleConsensusModelSphere<PointT>::computeModelCoefficients (
   model_coefficients[1] = 0.5f * m13 / m11;
   model_coefficients[2] = 0.5f * m14 / m11;
   // Radius
-  model_coefficients[3] = sqrtf (
-                                 model_coefficients[0] * model_coefficients[0] +
-                                 model_coefficients[1] * model_coefficients[1] +
-                                 model_coefficients[2] * model_coefficients[2] - m15 / m11);
+  model_coefficients[3] = std::sqrt (model_coefficients[0] * model_coefficients[0] +
+                                     model_coefficients[1] * model_coefficients[1] +
+                                     model_coefficients[2] * model_coefficients[2] - m15 / m11);
 
   return (true);
 }
@@ -135,7 +134,7 @@ pcl::SampleConsensusModelSphere<PointT>::getDistancesToModel (
   for (size_t i = 0; i < indices_->size (); ++i)
     // Calculate the distance from the point to the sphere as the difference between
     //dist(point,sphere_origin) and sphere_radius
-    distances[i] = fabs (sqrtf (
+    distances[i] = fabs (std::sqrt (
                                ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) *
                                ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) +
 
@@ -166,7 +165,7 @@ pcl::SampleConsensusModelSphere<PointT>::selectWithinDistance (
   // Iterate through the 3d points and calculate the distances from them to the sphere
   for (size_t i = 0; i < indices_->size (); ++i)
   {
-    double distance = fabs (sqrtf (
+    double distance = fabs (std::sqrt (
                           ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) *
                           ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) +
 
@@ -206,16 +205,16 @@ pcl::SampleConsensusModelSphere<PointT>::countWithinDistance (
   {
     // Calculate the distance from the point to the sphere as the difference between
     // dist(point,sphere_origin) and sphere_radius
-    if (fabs (sqrtf (
-                    ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) *
-                    ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) +
+    if (fabs (std::sqrt (
+                        ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) *
+                        ( input_->points[(*indices_)[i]].x - model_coefficients[0] ) +
 
-                    ( input_->points[(*indices_)[i]].y - model_coefficients[1] ) *
-                    ( input_->points[(*indices_)[i]].y - model_coefficients[1] ) +
+                        ( input_->points[(*indices_)[i]].y - model_coefficients[1] ) *
+                        ( input_->points[(*indices_)[i]].y - model_coefficients[1] ) +
 
-                    ( input_->points[(*indices_)[i]].z - model_coefficients[2] ) *
-                    ( input_->points[(*indices_)[i]].z - model_coefficients[2] )
-                    ) - model_coefficients[3]) < threshold)
+                        ( input_->points[(*indices_)[i]].z - model_coefficients[2] ) *
+                        ( input_->points[(*indices_)[i]].z - model_coefficients[2] )
+                        ) - model_coefficients[3]) < threshold)
       nr_p++;
   }
   return (nr_p);
